feat(agent): Upsert agent when logging in
