+++ app/controllers/api/v1/application_controller.rb
@@ -1,12 +1,11 @@
-class RdvSolidarites::InvalidSessionError < StandardError; end
-
 module Api
   module V1
     class ApplicationController < ActionController::Base
       skip_before_action :verify_authenticity_token
       respond_to :json
-      before_action :validate_session!
       include Pundit
+      include RdvSolidaritesSessionConcern
+      before_action :validate_session!
 
       private
 
@@ -16,18 +15,6 @@ module Api
 
       def current_agent
         @current_agent ||= Agent.find_by!(email: rdv_solidarites_session.uid)
-      end
-
-      def validate_session!
-        raise RdvSolidarites::InvalidSessionError unless rdv_solidarites_session.valid?
-      end
-
-      def rdv_solidarites_session
-        @rdv_solidarites_session ||= RdvSolidaritesSession.new(
-          uid: request.headers["uid"],
-          client: request.headers["client"],
-          access_token: request.headers["access-token"]
-        )
       end
 
       rescue_from RdvSolidarites::InvalidSessionError, with: :invalid_session
@@ -40,13 +27,6 @@ module Api
           json: {
             errors: ["Vous n'êtes pas autorisé à effectuer cette action"]
           }
-        )
-      end
-
-      def invalid_session
-        render(
-          json: { errors: ["Les identifiants de session RDV-Solidarités sont invalides"] },
-          status: :unauthorized
         )
       end
