+++ app/controllers/sessions_controller.rb
@@ -1,13 +1,15 @@
 class SessionsController < ApplicationController
-  PERMITTED_PARAMS = [:authenticity_token, :client, :uid, :access_token, { organisation_ids: [] }].freeze
   skip_before_action :authenticate_agent!, only: [:new, :create]
   wrap_parameters false
   respond_to :json, only: :create
 
+  include RdvSolidaritesSessionConcern
+  before_action :validate_session!, only: [:create]
+
   def new; end
 
   def create
-    return render_invalid_session unless rdv_solidarites_session_valid?
+    return render_cannot_retrieve_organisations unless retrieve_organisations.success?
 
     if find_or_create_agent.success?
       set_session_credentials
@@ -27,41 +29,31 @@ class SessionsController < ApplicationController
 
   def find_or_create_agent
     @find_or_create_agent ||= FindOrCreateAgent.call(
-      email: session_params[:uid], organisation_ids: session_params[:organisation_ids]
+      email: request.headers["uid"], organisation_ids: retrieve_organisations.organisations.map(&:id)
     )
   end
 
-  def created_agent
-    find_or_create_agent.agent
+  def render_cannot_retrieve_organisations
+    render json: { success: false, errors: retrieve_organisations.errors }, status: :unprocessable_entity
   end
 
   def set_session_credentials
-    session[:agent_id] = created_agent.id
+    session[:agent_id] = find_or_create_agent.agent.id
     session[:rdv_solidarites] = {
-      client: session_params[:client],
-      uid: session_params[:uid],
-      access_token: session_params[:access_token]
+      client: request.headers["client"],
+      uid: request.headers["uid"],
+      access_token: request.headers["access_token"]
     }
   end
 
-  def rdv_solidarites_session_valid?
-    RdvSolidaritesSession.new(
-      uid: session_params[:uid],
-      client: session_params[:client],
-      access_token: session_params[:access_token]
-    ).valid?
-  end
-
-  def render_invalid_session
-    render json: { success: false, errors: ["La session RDV-Solidarités n'est pas valide"] }
+  def retrieve_organisations
+    @retrieve_organisations ||= RdvSolidaritesApi::RetrieveOrganisations.call(
+      rdv_solidarites_session: rdv_solidarites_session
+    )
   end
 
   def clear_session
     session.delete(:agent_id)
     @current_agent = nil
-  end
-
-  def session_params
-    params.permit(*PERMITTED_PARAMS)
   end
 end
